refactor(ui): リネーム
