+++ src/view/index/filter/FilterByPartsOffCanvas.svelte
